requirements.txt
asttokens==2.4.1
base==0.0.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
comm==0.2.0
cryptography==41.0.5
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
flask==3.0.0
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter-client==8.6.0
jupyter-core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.2
parso==0.8.3
platformdirs==4.0.0
postgres==4.0
prompt-toolkit==3.0.40
psutil==5.9.6
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing-extensions==4.8.0
tzdata==2023.3
validate-email-address==1
wcwidth==0.2.9
werkzeug==3.0.1
zipp==3.17.0
